fix pollution of :contains pseudo selector (#6434)
* fix overriding :contains pseudo selector

* dont inject livereload inside spec iframe...

* add issue comment
